OCPBUGS-59750: Move the legacy container failures
Move the legacy container failures monitor tests to a dedicated
MonitorTest implementation and removed the test from the default suites
for disruptive test suites.
