+++ pkg/monitortests/kubelet/containerfailures/exclusions.go
@@ -1,4 +1,4 @@
-package legacynodemonitortests
+package containerfailures
 
 import (
 	"regexp"
