+++ pkg/monitortests/kubelet/containerfailures/exclusions_test.go
@@ -1,4 +1,4 @@
-package legacynodemonitortests
+package containerfailures
 
 import (
 	"testing"
