+++ pkg/monitortests/node/legacynodemonitortests/monitortest.go
@@ -41,13 +41,7 @@ func (*legacyMonitorTests) ConstructComputedIntervals(ctx context.Context, start
 func (w *legacyMonitorTests) EvaluateTestsFromConstructedIntervals(ctx context.Context, finalIntervals monitorapi.Intervals) ([]*junitapi.JUnitTestCase, error) {
 
 	clusterData, _ := platformidentification.BuildClusterData(context.Background(), w.adminRESTConfig)
-
-	containerFailures, err := testContainerFailures(w.adminRESTConfig, finalIntervals)
-	if err != nil {
-		return nil, err
-	}
-	junits := []*junitapi.JUnitTestCase{}
-	junits = append(junits, containerFailures...)
+	var junits []*junitapi.JUnitTestCase
 	junits = append(junits, testDeleteGracePeriodZero(finalIntervals)...)
 	junits = append(junits, testKubeApiserverProcessOverlap(finalIntervals)...)
 	junits = append(junits, testKubeAPIServerGracefulTermination(finalIntervals)...)
